Issue #IQ-149 fix: package renamed
