+++ publish-pipeline/questionset-republish/src/main/scala/org/sunbird/job/questionset/republish/domain/Models.scala
@@ -1,3 +1,3 @@
-package org.sunbird.job.questionset.publish.domain
+package org.sunbird.job.questionset.republish.domain
 
 case class PublishMetadata(identifier: String, objectType: String, mimeType: String, pkgVersion: Double, publishType: String, lastPublishedBy: String)
