+++ publish-pipeline/questionset-republish/src/main/scala/org/sunbird/job/questionset/republish/helpers/QuestionSetPublisher.scala
@@ -1,4 +1,4 @@
-package org.sunbird.job.questionset.publish.helpers
+package org.sunbird.job.questionset.republish.helpers
 
 import java.io.File
