+++ publish-pipeline/questionset-republish/src/main/scala/org/sunbird/job/questionset/republish/task/QuestionSetRePublishConfig.scala
@@ -1,11 +1,11 @@
-package org.sunbird.job.questionset.task
+package org.sunbird.job.questionset.republish.task
 
 import com.typesafe.config.Config
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.java.typeutils.TypeExtractor
 import org.apache.flink.streaming.api.scala.OutputTag
 import org.sunbird.job.publish.config.PublishConfig
-import org.sunbird.job.questionset.publish.domain.PublishMetadata
+import org.sunbird.job.questionset.republish.domain.PublishMetadata
 
 import java.util
 import scala.collection.JavaConverters._
