+++ publish-pipeline/questionset-republish/src/test/scala/org/sunbird/job/republish/helpers/spec/QuestionSetPublisherSpec.scala
@@ -1,4 +1,4 @@
-package org.sunbird.job.publish.helpers.spec
+package org.sunbird.job.republish.helpers.spec
 
 import com.typesafe.config.{Config, ConfigFactory}
 import org.cassandraunit.CQLDataLoader
@@ -9,8 +9,8 @@ import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
 import org.scalatestplus.mockito.MockitoSugar
 import org.sunbird.job.domain.`object`.DefinitionCache
 import org.sunbird.job.publish.core.{DefinitionConfig, ExtDataConfig, ObjectData, ObjectExtData}
-import org.sunbird.job.questionset.publish.helpers.QuestionSetPublisher
-import org.sunbird.job.questionset.task.QuestionSetRePublishConfig
+import org.sunbird.job.questionset.republish.helpers.QuestionSetPublisher
+import org.sunbird.job.questionset.republish.task.QuestionSetRePublishConfig
 import org.sunbird.job.util.{CassandraUtil, CloudStorageUtil, Neo4JUtil}
 
 class QuestionSetPublisherSpec extends FlatSpec with BeforeAndAfterAll with Matchers with MockitoSugar {
